requirements.txt
langgraph
langchain-core
langchain_community
langchain_google_genai
langchain_openai
langchain_chroma
chromadb
hubspot-api-client
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
colorama
python-dotenv
bs4
unstructured
linkedin-api
supabase
requests
html2text